+++ src/main/java/kr/co/khedu/fitroutine/post/controller/PostController.java
@@ -9,7 +9,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.net.URI;
 import java.util.List;
@@ -70,37 +69,6 @@ public class PostController {
     @DeleteMapping("/posts/{postId}")
     public ResponseEntity<Void> deletePost(@PathVariable long postId) {
         postService.deletePost(postId);
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/posts/{postId}/images")
-    public ResponseEntity<List<? extends ImageResponse>> getImages(@PathVariable long postId) {
-        return ResponseEntity.ok(postService.getImages(postId));
-    }
-
-    @GetMapping("/images/{imageId}")
-    public ResponseEntity<ImageResponse> getImage(@PathVariable long imageId) {
-        return ResponseEntity.ok(postService.getImage(imageId));
-    }
-
-    @PreAuthorize("@postService.isPostOwner(#postId, principal)")
-    @PostMapping("/posts/{postId}/images")
-    public ResponseEntity<ImageResponse> createImage(
-            @PathVariable long postId,
-            @RequestPart MultipartFile multipartFile
-    ) {
-        return ResponseEntity.ok(
-                postService.createImage(
-                        postId,
-                        postService.toCreateRequest(multipartFile)
-                )
-        );
-    }
-
-    @PreAuthorize("@postService.isImageOwner(#imageId, principal)")
-    @DeleteMapping("/images/{imageId}")
-    public ResponseEntity<Void> deleteImage(@PathVariable long imageId) {
-        postService.deleteImage(imageId);
         return ResponseEntity.noContent().build();
     }
 }
