+++ src/main/java/kr/co/khedu/fitroutine/post/service/PostService.java
@@ -3,31 +3,19 @@ package kr.co.khedu.fitroutine.post.service;
 import kr.co.khedu.fitroutine.post.mapper.PostMapper;
 import kr.co.khedu.fitroutine.post.model.dto.*;
 import kr.co.khedu.fitroutine.security.model.dto.UserDetailsImpl;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.FilenameUtils;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.UUID;
 
 @Service
 @Transactional
 public class PostService {
     private final PostMapper postMapper;
-    private final String uploadDir;
 
-    public PostService(
-            PostMapper postMapper,
-            @Value("${file.upload-dir}") String uploadDir
-    ) {
+    public PostService(PostMapper postMapper) {
         this.postMapper = postMapper;
-        this.uploadDir = uploadDir;
     }
 
     @Transactional(readOnly = true)
@@ -75,67 +63,6 @@ public class PostService {
     public void deletePost(long postId) {
         if (postMapper.deletePost(postId) != 1) {
             throw new NoSuchElementException("포스트가 존재하지 않습니다. id=" + postId);
-        }
-    }
-
-    @Transactional(readOnly = true)
-    public List<? extends ImageResponse> getImages(long postId) {
-        return postMapper.selectImagesByPostId(postId);
-    }
-
-    @Transactional(readOnly = true)
-    public ImageResponse getImage(long imageId) {
-        ImageResponse imageResponse = postMapper.selectImageById(imageId);
-        if (imageResponse == null) {
-            throw new NoSuchElementException("이미지를 찾을 수 없습니다. id=" + imageId);
-        }
-
-        return imageResponse;
-    }
-
-    @Transactional(readOnly = true)
-    public boolean isImageOwner(long imageId, UserDetailsImpl userDetails) {
-        return postMapper.existsImageByMemberId(
-                imageId,
-                userDetails.getMemberId()
-        ) == 1;
-    }
-
-    public ImageResponse createImage(long postId, ImageCreateRequest createRequest) {
-        postMapper.insertImage(postId, createRequest);
-
-        Long imageId = createRequest.getImageId();
-        if (imageId == null) {
-            throw new IllegalStateException("파일 행을 추가할 수 없습니다.");
-        }
-
-        return getImage(createRequest.getImageId());
-    }
-
-    public ImageCreateRequest toCreateRequest(MultipartFile multipartFile) {
-        if (multipartFile.getOriginalFilename() == null) {
-            throw new IllegalStateException("파일 이름이 유효하지 않습니다.");
-        }
-
-        String originName = multipartFile.getOriginalFilename();
-        String changeName = UUID.randomUUID() + "." + FilenameUtils.getExtension(originName);
-
-        File destination = new File(uploadDir, changeName);
-        try {
-            FileUtils.copyInputStreamToFile(multipartFile.getInputStream(), destination);
-        } catch (IOException exception) {
-            throw new RuntimeException("파일 저장에 실패했습니다.", exception);
-        }
-
-        return ImageCreateRequest.builder()
-                .originName(originName)
-                .changeName(changeName)
-                .build();
-    }
-
-    public void deleteImage(long imageId) {
-        if (postMapper.deleteImage(imageId) != 1) {
-            throw new NoSuchElementException("이미지가 존재하지 않습니다. id=" + imageId);
         }
     }
 }
