+++ src/main/resources/mapper/post-image-mapper.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//KO"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="kr.co.khedu.fitroutine.post.mapper.PostImageMapper">
+    <select id="selectImagesByPostId" resultType="kr.co.khedu.fitroutine.post.model.dto.ImageResponse">
+        SELECT
+            IMAGE_ID,
+            ORIGIN_NAME,
+            CHANGE_NAME,
+            POST_ID
+        FROM TB_IMAGE
+        WHERE POST_ID = #{postId}
+    </select>
+    <select id="selectImageById" resultType="kr.co.khedu.fitroutine.post.model.dto.ImageResponse">
+        SELECT
+            IMAGE_ID,
+            ORIGIN_NAME,
+            CHANGE_NAME,
+            POST_ID
+        FROM TB_IMAGE
+        WHERE IMAGE_ID = #{imageId}
+    </select>
+    <select id="existsImageByMemberId" resultType="int">
+        SELECT
+            CASE
+                WHEN EXISTS (
+                    SELECT 1
+                    FROM TB_MEMBER M
+                    JOIN TB_BLOG B ON B.MEMBER_ID = M.MEMBER_ID
+                    JOIN TB_POST P ON P.BLOG_ID = B.BLOG_ID
+                    JOIN TB_IMAGE I ON I.POST_ID = P.POST_ID
+                    WHERE M.MEMBER_ID = #{memberId}
+                    AND I.IMAGE_ID = #{imageId}
+                )
+                THEN 1
+                ELSE 0
+            END
+        FROM DUAL
+    </select>
+    <insert id="insertImage">
+        <selectKey keyProperty="createRequest.imageId" resultType="long" order="BEFORE">
+            SELECT SEQ_IMAGE_ID.NEXTVAL
+            FROM dual
+        </selectKey>
+        INSERT INTO TB_IMAGE (
+            IMAGE_ID,
+            ORIGIN_NAME,
+            CHANGE_NAME,
+            POST_ID
+        )
+        VALUES (
+            #{createRequest.imageId},
+            #{createRequest.originName},
+            #{createRequest.changeName},
+            #{postId}
+        )
+    </insert>
+    <delete id="deleteImage">
+        DELETE FROM TB_IMAGE
+        WHERE IMAGE_ID = #{imageId}
+    </delete>
+</mapper>
